23.01.25 여원영 Bbs & Hot / sevice inter 등 rename 작업
