+++ src/main/java/com/jav4/acomm/bbs/BbsService.java
@@ -5,64 +5,66 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.jav4.acomm.apt.AptVO;
+import com.jav4.acomm.member.MemberVO;
+
 @Service
 public class BbsService implements BbsServiceInter {
 
 	@Autowired
-	BbsHotDAOInter dao;
+	BbsDAOInter dao;
 	
-	// 인기게시글 5개만 검색
+	// apt * by apt_code
 	@Override
-	public List<BbsVO> list(BbsVO vo) {
-		 return dao.list(vo);
+	public AptVO aptAll(AptVO avo) {
+		 return dao.aptAll(avo);
 	}
 	
-	// 인기게시글 전체 검색
+	// member * by member_code
 	@Override
-	public List<BbsVO> listAll(BbsVO vo) {
-		return dao.listAll(vo);
+	public MemberVO memberAll(MemberVO mvo) {
+		 return dao.memberAll(mvo);
 	}
 	
-	// 조회수
+	// bbs의 전체 게시판 보기
 	@Override
-	public void bbsCount(int bbs_id) {
-		dao.bbsCount(bbs_id);
+	public List<BbsVO> getList5(BbsVO vo) {
+		 return dao.getList5(vo);
 	}
 	
-	// 하나 상세 검색
+	// cate별로 상세 게시판 보기
 	@Override
-	public BbsVO one(BbsVO vo) {
-		return dao.one(vo);
+	public List<BbsVO> getListCate(BbsVO bvo) {
+		 return dao.getListCate(bvo);
 	}
 	
-//	// bbs 좋아요
-//	public void bbslike(LikeVO vo) {
-//		dao.bbshotlike(vo);
-//	}
-	
-	// 통합 테이블 좋아요 추가
+	//
 	@Override
-	public void bbslike(BbslikeVO vo) {
-			dao.bbslike(vo);
+	public MemberVO id2cls(MemberVO mvo) {
+		 return dao.id2cls(mvo);
 	}
 	
-	// 통합 테이블 좋아요 취소
+	//
 	@Override
-	public void bbslikedel(BbslikeVO vo) {
-//		BbslikeVO one = dao.bbslikeone(vo);
-		dao.bbslikedel(vo);
-//		return one;
+	public BbsVO getBbsPost(BbsVO vo) {
+		 return dao.getBbsPost(vo);
 	}
 	
-	// bbs_id와 member_code가 일치하는 테이블 전체검색
+	//
 	@Override
-	public List<BbslikeVO> hotlist(BbslikeVO vo) {
-		 return dao.hotlist(vo);
+	public void insertPost(BbsVO vo) {
+		 dao.insertPost(vo);
 	}
 	
-	// bbs_id와 member_code가 일치하는 테이블 하나검색
+	//
 	@Override
-	public BbslikeVO bbslikeone(BbslikeVO vo) {
-		return dao.bbslikeone(vo);
+	public void updatePost(BbsVO vo) {
+		 dao.updatePost(vo);
 	}
+	//
+	@Override
+	public void deletePost(BbsVO vo) {
+		 dao.deletePost(vo);
+	}
+	
 }
